+++ spring-addons-starter-rest/src/main/java/com/c4_soft/springaddons/rest/reactive/ServerWebClientFactoryBean.java
@@ -4,10 +4,9 @@ import java.util.Optional;
 import org.jspecify.annotations.Nullable;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.FactoryBean;
-import org.springframework.boot.http.client.reactive.ClientHttpConnectorSettings;
-import org.springframework.boot.http.client.reactive.ClientHttpConnectorBuilder;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
+import org.springframework.http.client.reactive.ClientHttpConnector;
 import org.springframework.security.oauth2.client.ReactiveOAuth2AuthorizedClientManager;
 import org.springframework.security.oauth2.client.ReactiveOAuth2AuthorizedClientService;
 import org.springframework.security.oauth2.client.web.server.ServerOAuth2AuthorizedClientRepository;
@@ -29,8 +28,7 @@ public class ServerWebClientFactoryBean implements FactoryBean<WebClient>, Appli
       Optional.empty();
   private Optional<ReactiveOAuth2AuthorizedClientService> authorizedClientService =
       Optional.empty();
-  private Optional<ClientHttpConnectorBuilder<?>> clientHttpConnectorBuilder;
-  private Optional<ClientHttpConnectorSettings> httpClientSettings;
+  private Optional<ClientHttpConnector> clientHttpConnector = Optional.empty();
   private WebClient.Builder webClientBuilder;
   private @Nullable ApplicationContext applicationContext;
 
@@ -52,8 +50,7 @@ public class ServerWebClientFactoryBean implements FactoryBean<WebClient>, Appli
     builderFactoryBean.setAuthorizedClientManager(authorizedClientManager);
     builderFactoryBean.setAuthorizedClientRepository(authorizedClientRepository);
     builderFactoryBean.setAuthorizedClientService(authorizedClientService);
-    builderFactoryBean.setClientHttpConnectorBuilder(clientHttpConnectorBuilder);
-    builderFactoryBean.setHttpClientSettings(httpClientSettings);
+    builderFactoryBean.setClientHttpConnector(clientHttpConnector);
     builderFactoryBean.setWebClientBuilder(webClientBuilder);
     return Optional.ofNullable(builderFactoryBean.getObject()).map(WebClient.Builder::build)
         .orElse(null);